+++ sys/src/games/doom/p_saveg.c
@@ -37,9 +37,7 @@ byte*		save_p;
 
 // Pads save_p to a 4-byte boundary
 //  so that the load/save works on SGI&Gecko.
-#define PADSAVEP()	save_p += (4 - ((int) save_p & 3)) & 3
-
-
+#define PADSAVEP()	save_p += (4 - ((uintptr) save_p & 3)) & 3
 
 //
 // P_ArchivePlayers
@@ -101,7 +99,7 @@ void P_UnArchivePlayers (void)
 	    if (players[i]. psprites[j].state)
 	    {
 		players[i]. psprites[j].state 
-		    = &states[ (int)players[i].psprites[j].state ];
+		    = &states[ (int)(uintptr)players[i].psprites[j].state ];
 	    }
 	}
     }
@@ -299,11 +297,11 @@ void P_UnArchiveThinkers (void)
 	    mobj = Z_Malloc (sizeof(*mobj), PU_LEVEL, NULL);
 	    memcpy (mobj, save_p, sizeof(*mobj));
 	    save_p += sizeof(*mobj);
-	    mobj->state = &states[(int)mobj->state];
+	    mobj->state = &states[(int)(uintptr)mobj->state];
 	    mobj->target = NULL;
 	    if (mobj->player)
 	    {
-		mobj->player = &players[(int)mobj->player-1];
+		mobj->player = &players[(int)(uintptr)mobj->player-1];
 		mobj->player->mo = mobj;
 	    }
 	    P_SetThingPosition (mobj);
@@ -498,7 +496,7 @@ void P_UnArchiveSpecials (void)
 	    ceiling = Z_Malloc (sizeof(*ceiling), PU_LEVEL, NULL);
 	    memcpy (ceiling, save_p, sizeof(*ceiling));
 	    save_p += sizeof(*ceiling);
-	    ceiling->sector = &sectors[(int)ceiling->sector];
+	    ceiling->sector = &sectors[(int)(uintptr)ceiling->sector];
 	    ceiling->sector->specialdata = ceiling;
 
 	    if (ceiling->thinker.function)
@@ -513,7 +511,7 @@ void P_UnArchiveSpecials (void)
 	    door = Z_Malloc (sizeof(*door), PU_LEVEL, NULL);
 	    memcpy (door, save_p, sizeof(*door));
 	    save_p += sizeof(*door);
-	    door->sector = &sectors[(int)door->sector];
+	    door->sector = &sectors[(int)(uintptr)door->sector];
 	    door->sector->specialdata = door;
 	    door->thinker.function = T_VerticalDoor;
 	    P_AddThinker (&door->thinker);
@@ -524,7 +522,7 @@ void P_UnArchiveSpecials (void)
 	    floor = Z_Malloc (sizeof(*floor), PU_LEVEL, NULL);
 	    memcpy (floor, save_p, sizeof(*floor));
 	    save_p += sizeof(*floor);
-	    floor->sector = &sectors[(int)floor->sector];
+	    floor->sector = &sectors[(int)(uintptr)floor->sector];
 	    floor->sector->specialdata = floor;
 	    floor->thinker.function = T_MoveFloor;
 	    P_AddThinker (&floor->thinker);
@@ -535,7 +533,7 @@ void P_UnArchiveSpecials (void)
 	    plat = Z_Malloc (sizeof(*plat), PU_LEVEL, NULL);
 	    memcpy (plat, save_p, sizeof(*plat));
 	    save_p += sizeof(*plat);
-	    plat->sector = &sectors[(int)plat->sector];
+	    plat->sector = &sectors[(int)(uintptr)plat->sector];
 	    plat->sector->specialdata = plat;
 
 	    if (plat->thinker.function)
@@ -550,7 +548,7 @@ void P_UnArchiveSpecials (void)
 	    flash = Z_Malloc (sizeof(*flash), PU_LEVEL, NULL);
 	    memcpy (flash, save_p, sizeof(*flash));
 	    save_p += sizeof(*flash);
-	    flash->sector = &sectors[(int)flash->sector];
+	    flash->sector = &sectors[(int)(uintptr)flash->sector];
 	    flash->thinker.function = T_LightFlash;
 	    P_AddThinker (&flash->thinker);
 	    break;
@@ -560,7 +558,7 @@ void P_UnArchiveSpecials (void)
 	    strobe = Z_Malloc (sizeof(*strobe), PU_LEVEL, NULL);
 	    memcpy (strobe, save_p, sizeof(*strobe));
 	    save_p += sizeof(*strobe);
-	    strobe->sector = &sectors[(int)strobe->sector];
+	    strobe->sector = &sectors[(int)(uintptr)strobe->sector];
 	    strobe->thinker.function = T_StrobeFlash;
 	    P_AddThinker (&strobe->thinker);
 	    break;
@@ -570,7 +568,7 @@ void P_UnArchiveSpecials (void)
 	    glow = Z_Malloc (sizeof(*glow), PU_LEVEL, NULL);
 	    memcpy (glow, save_p, sizeof(*glow));
 	    save_p += sizeof(*glow);
-	    glow->sector = &sectors[(int)glow->sector];
+	    glow->sector = &sectors[(int)(uintptr)glow->sector];
 	    glow->thinker.function = T_Glow;
 	    P_AddThinker (&glow->thinker);
 	    break;
